Add an option deleteOnError to delete erroneous bagFiles
You can pass the option deleteOnError to either of the launchfiles
analyse_bag_file and analyse_remaining_bag_files. Valid options include
'ask', 'yes', 'no'.
If you pass ask, the user is prompted to specify if he wants to keep or
delete the bag file. If he passes yes or no to either the argument or
the question the bag file is deleted/kept respecitvely.